Refresh tableView data only after a location has been found
Obviously only if sorting is set to distance sorting. The upside here
is that the user doesn't see a default sorted list when refreshing and
being set to distance sorting. Sometimes iOS takes a few seconds before
finding a fix and that can look stupid.

I tested this with a few cases of no internet connection, changing
settings and revoking location permissions. It's a bit wonky sometimes
and not too perfect, but I guess fine for normal use. In extreme cases
it will take a bit longer to function normally, but that would
definitely be the user's fault! :D

Closes #45