Add new error case for broken server data
This being what happened a few days ago (conveniently when demoing the
app to someone who also happened to want to take photos...), it might
be a good idea to catch this. If any sections lotcount is now 0, an
alert is displayed. I don't believe that's a case that can happen
normally.

Closes #43