Fix UIRefreshControl display bug
At least I hope so. Sometimes there would be additional whitespace at
the top of the table view where the refreshcontrol would only start
below that. Not really reproduceable, but this seems like a logical
solution for this. Let's hope it doesn't appear anymore.

See http://stackoverflow.com/questions/19240915/fix-uitableviewcontroller-offset-due-to-uirefreshcontrol-in-ios-7