Another fix for UIRefreshControl bug
I'm guessing it was caused by updateData() being called from the
viewDidLoad(). That attributed string stuff for the UIRefreshControl is
the bad guy as far as I can tell which shouldn't be called from the
viewDidLoad() for whatever reason. Since the update is now being run
by the app delegate that shouldn't happen, but I've got to cheat with
the activity indicator. Let's hope this is the final fix for that bug.