Add PrognosisViewController
Accessible via the settings, this experimental feature shows a prognosis
for the availability of spots in the parking lot Centrum Galerie. The
user can choose a date via the date picker which is then looked up in
the CSV prognosis data from @balzer82.

Also uses naoty/SwiftCSV for parsing the CSV data. Albeit slightly
modified with a custom initializer via a String instead of NSURL.

See #47