Fix some searchbar inconsistencies
It's now set to inactive on segue to mapVC, but my UX feeling says that
this shouldn't be happening, as a user's search result is typically not
removed when coming back from a detailvc.

Also only scroll past searchbar if tableview is at the top. But all of
this is rather unclean as well. I think other apps just show their
search bar... But every option looks so wrong here :/

See #40