Lots are shown again with the new backend 🎉
Lots can be displayed for each city that's supported. Only
problem is that the server is sending the nodata state fairly
often, which means I can no longer use that state to show a ?
for a lot.

Also the handling of the supported cities is a bit weird,
maybe I'll think of a better way to persist that data in the
future.

See #70